06-lab-complete/production/app/requirements.txt
streamlit==1.45.1
openai==1.76.0
python-dotenv==1.0.1
pytest==8.3.5
